+++ container-setup/install/install-kustomize.sh
@@ -8,13 +8,18 @@ fi
 echo "in container";
 source /container-setup/install/helpers.sh
 
-# export kustomizeversion=v3@v3.5.4
 
 mkdir /usr/local/kustomize;
 pushd /usr/local/kustomize;
 
-remove_coloring go mod init tmp;
-remove_coloring go get -v -u sigs.k8s.io/kustomize/kustomize/${kustomizeversion};
+remove_coloring go mod init tmp
+remove_coloring go install sigs.k8s.io/kustomize/kustomize/v3
+if [[ ! -f /root/go/bin/kustomize ]]
+then 
+	echo "the binary was not in the GOBIN, exiting"
+	exit 1
+fi
+
 ln -s /root/go/bin/kustomize /usr/local/bin/kustomize;
 
 popd
